+++ ModLoader/ModEntry.cs
@@ -173,7 +173,8 @@ namespace Sheepy.Modnix {
 
       private Version GetVersion ( object target ) {
          if ( LowerAndIsEmpty( target, out string id ) ) lock ( Metadata ) return Metadata.Version;
-         return ModScanner.GetVersionById( id );
+         ModScanner.GetVersionById( id, out _, out Version ver );
+         return ver;
       }
 
       private ModEntry GetMod ( object target ) {
@@ -526,6 +527,7 @@ namespace Sheepy.Modnix {
       public TextSet   Contact;
       public TextSet   Copyright;
 
+      public AppVer[]  Avoids;
       public AppVer[]  Requires;
       public AppVer[]  Disables;
       public long      LoadIndex;
@@ -555,6 +557,7 @@ namespace Sheepy.Modnix {
             CopyNonNull( overrider.Url, ref Url );
             CopyNonNull( overrider.Contact, ref Contact );
             CopyNonNull( overrider.Copyright, ref Copyright );
+            CopyNonNull( overrider.Avoids, ref Avoids );
             CopyNonNull( overrider.Requires, ref Requires );
             CopyNonNull( overrider.Disables, ref Disables );
             CopyNonNull( overrider.LoadIndex, ref LoadIndex );
@@ -591,6 +594,7 @@ namespace Sheepy.Modnix {
          NormTextSet( ref Url );
          NormTextSet( ref Contact );
          NormTextSet( ref Copyright );
+         NormAppVer( ref Avoids );
          NormAppVer( ref Requires );
          NormAppVer( ref Disables );
          NormStringArray( ref Mods );