+++ MainGUI/Resources/Changelog.md
@@ -2,7 +2,10 @@ Modnix Changelog
 
 # Version 2.4, in development
 
+* New: Avoids field of mod_info - self-kill if target exists.
 * Fix: mod_info with Version field are now parsed correctly.
+* Fix: A mod's requirement(s) on itself are now ignored.
+* Fix: LoadIndex and LogLevel may be overrode together in modnix config, instead of index over level.
 * Mod: "assemblies" api now returns null when mod is not found.
 * Mod: api( "assemblies", "modnix" ) no longer returns ppml assembly.
 * Mod: Loader log now tries to capture game crashs originating from the scope of CLR.
@@ -10,7 +13,8 @@ Modnix Changelog
 * Mod: Loader now logs mscorlib version and end of phases at verbose level.
 * Mod: Requires and Disables are now resolved together in each resolve loop, instead of only Requires.
 * Mod: Multiple requires on the same mod are now processed as "or" instead of "and".
-* Mod: Flags and Actions in mod_info (intended for Modnix 3) now leave a notice in Manager and log.
+* Mod: Flags and Actions in mod_info (intended for Modnix 3) now trigger a notice in Manager and log.
+* Mod: A mod that try to mod_info.Disables itself now triggers a warning, same for Avoids/Requires.
 
 # Version 2.3.1, 2020-05-03